third_party/WebKit/LayoutTests/fast/dom/shadow/closed-mode-deep-combinators-expected.txt

CONSOLE WARNING: /deep/ combinator is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
PASS result.length is 5
PASS node.id is "openhost"
PASS node.id is "openhost_in_openshadow"
PASS node.id is "div_open_in_open"
PASS node.id is "closedhost_in_openshadow"
PASS node.id is "closedhost"
PASS successfullyParsed is true

TEST COMPLETE
